requirements.txt
web3==7.14.0
python-dotenv==1.1.1
pyTelegramBotAPI==4.29.1
hexbytes==1.3.1
eth_abi==5.2.0
numpy==2.3.3
requests==2.32.3
python-telegram-bot==22.5
fastapi==0.128.0
uvicorn[standard]
pymongo==4.16.0
PyJWT==2.10.1
privy-client==0.5.0